Add files via upload
Developing a simple budget tracking 
application to encourage responsible saving habits. This app will be 
promoted to bank customers of all ages and technical backgrounds.
• Log in with a username and password. 
• Access a menu system that allows the user to add income, add expense entries, view current 
balance and recent transactions, and set up a simple savings goal. 
• A savings overview that shows the current balance after subtracting the expenses from income. 
This could display how much is potentially available for saving. 
• A way of displaying the spending habits of the user either according to the categories, or the 
otherwise. This can be displayed as a percentage or some other way.